+++ kipper/core/src/compiler/target-presets/compile-target.ts
@@ -1,33 +1,36 @@
 /**
- * Compilation translation class specifying how a Kipper parse tree shall be translated into a specific language.
+ * Compilation translation class specifying how a translation target should be implemented.
  * @author Luna Klatzer
  * @copyright 2021-2022 Luna Klatzer
- * @since 0.5.0
+ * @since 0.10.0
  */
 
 import type { KipperTargetBuiltInGenerator, KipperTargetCodeGenerator } from "./translation";
-import type { KipperTargetSemanticAnalyser } from "./semantics";
+import type { KipperTargetSemanticAnalyser } from "./target-semantic-analyser";
 
 /**
  * Represents a Kipper compilation translation defining how a Kipper parse tree shall
  * be translated into a specific language.
- * @since 0.5.0
+ * @since 0.10.0
  */
-export class KipperCompileTarget {
+export abstract class KipperCompileTarget {
 	public readonly targetName: string;
 	public readonly semanticAnalyser: KipperTargetSemanticAnalyser;
 	public readonly codeGenerator: KipperTargetCodeGenerator;
 	public readonly builtInGenerator: KipperTargetBuiltInGenerator;
+	public readonly fileExtension: string;
 
-	constructor(
+	protected constructor(
 		targetName: string,
 		semanticAnalyser: KipperTargetSemanticAnalyser,
 		codeGenerator: KipperTargetCodeGenerator,
 		builtInGenerator: KipperTargetBuiltInGenerator,
+		fileExtension: string,
 	) {
 		this.targetName = targetName;
 		this.semanticAnalyser = semanticAnalyser;
 		this.codeGenerator = codeGenerator;
 		this.builtInGenerator = builtInGenerator;
+		this.fileExtension = fileExtension;
 	}
 }