+++ kipper/core/src/compiler/target-presets/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Presets for a Kipper translation target.
+ * @author Luna Klatzer
+ * @copyright 2021-2022 Luna Klatzer
+ * @since 0.10.0
+ */
+
+export * from "./translation";
+export * from "./compile-target";
+export * from "./target-semantic-analyser";