+++ kipper/target-ts/src/semantic-analyser.ts
@@ -1,0 +1,15 @@
+/**
+ * The TypeScript target-specific semantic analyser.
+ * @author Luna Klatzer
+ * @copyright 2021-2022 Luna Klatzer
+ * @since 0.8.0
+ */
+import { JavaScriptTargetSemanticAnalyser } from "@kipper/target-js";
+
+/**
+ * The TypeScript target-specific semantic analyser.
+ * @since 0.8.0
+ */
+export class TypeScriptTargetSemanticAnalyser extends JavaScriptTargetSemanticAnalyser {}
+
+export { reservedIdentifiers } from "@kipper/target-js";